--- src/main/java/com/kdt/team04/domain/match/review/entity/MatchRecordValue.java
@@ -1,5 +1,0 @@
-package com.kdt.team04.domain.match.review.entity;
-
-public enum MatchRecordValue {
-	WIN, LOSE, DRAW
-}
