--- src/main/java/com/kdt/team04/domain/match/review/repository/MatchRecordRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.kdt.team04.domain.match.review.repository;
-
-import com.kdt.team04.domain.match.review.dto.MatchRecordResponse;
-
-public interface MatchRecordRepositoryCustom {
-	MatchRecordResponse.TotalCount getTeamTotalCount(Long teamId);
-}
