+++ src/main/java/com/kdt/team04/domain/matches/match/entity/MatchStatus.java
@@ -1,4 +1,4 @@
-package com.kdt.team04.domain.match.post.entity;
+package com.kdt.team04.domain.matches.match.entity;
 
 public enum MatchStatus {
 	WAITING, IN_GAME, END
