+++ src/main/java/com/kdt/team04/domain/matches/match/entity/MatchType.java
@@ -1,4 +1,4 @@
-package com.kdt.team04.domain.match.post.entity;
+package com.kdt.team04.domain.matches.match.entity;
 
 public enum MatchType {
 	TEAM_MATCH, INDIVIDUAL_MATCH
