+++ src/main/java/com/kdt/team04/domain/matches/request/entity/MatchRequestStatus.java
@@ -1,4 +1,4 @@
-package com.kdt.team04.domain.match.request.entity;
+package com.kdt.team04.domain.matches.request.entity;
 
 public enum MatchRequestStatus {
 	WAITING, APPROVED
