+++ src/main/java/com/kdt/team04/domain/matches/review/dto/MatchRecordResponse.java
@@ -1,4 +1,4 @@
-package com.kdt.team04.domain.match.review.dto;
+package com.kdt.team04.domain.matches.review.dto;
 
 public record MatchRecordResponse() {
 	public record TotalCount(int win, int draw, int lose) {
