+++ src/main/java/com/kdt/team04/domain/matches/review/dto/MatchReviewResponse.java
@@ -1,4 +1,4 @@
-package com.kdt.team04.domain.match.review.dto;
+package com.kdt.team04.domain.matches.review.dto;
 
 public record MatchReviewResponse(){
 	public record TotalCount(int bestCount, int likeCount, int dislikeCount){
