+++ src/main/java/com/kdt/team04/domain/matches/review/entity/MatchRecordValue.java
@@ -1,0 +1,5 @@
+package com.kdt.team04.domain.matches.review.entity;
+
+public enum MatchRecordValue {
+	WIN, LOSE, DRAW
+}
