+++ src/main/java/com/kdt/team04/domain/matches/review/entity/MatchReviewValue.java
@@ -1,4 +1,4 @@
-package com.kdt.team04.domain.match.review.entity;
+package com.kdt.team04.domain.matches.review.entity;
 
 public enum MatchReviewValue {
 	BEST, LIKE, DISLIKE
