+++ src/main/java/com/kdt/team04/domain/matches/review/repository/MatchRecordRepository.java
@@ -1,8 +1,8 @@
-package com.kdt.team04.domain.match.review.repository;
+package com.kdt.team04.domain.matches.review.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.kdt.team04.domain.match.review.entity.MatchRecord;
+import com.kdt.team04.domain.matches.review.entity.MatchRecord;
 
 public interface MatchRecordRepository extends JpaRepository<MatchRecord, Long>, MatchRecordRepositoryCustom {
 }
