+++ src/main/java/com/kdt/team04/domain/matches/review/repository/MatchRecordRepositoryImpl.java
@@ -1,11 +1,11 @@
-package com.kdt.team04.domain.match.review.repository;
+package com.kdt.team04.domain.matches.review.repository;
 
-import static com.kdt.team04.domain.match.review.entity.QMatchRecord.matchRecord;
+import static com.kdt.team04.domain.matches.review.entity.QMatchRecord.matchRecord;
 
 import org.springframework.stereotype.Repository;
 
-import com.kdt.team04.domain.match.review.dto.MatchRecordResponse;
-import com.kdt.team04.domain.match.review.entity.MatchRecordValue;
+import com.kdt.team04.domain.matches.review.dto.MatchRecordResponse;
+import com.kdt.team04.domain.matches.review.entity.MatchRecordValue;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
