+++ src/main/java/com/kdt/team04/domain/matches/review/service/MatchRecordGiverService.java
@@ -1,9 +1,9 @@
-package com.kdt.team04.domain.match.review.service;
+package com.kdt.team04.domain.matches.review.service;
 
 import org.springframework.stereotype.Service;
 
-import com.kdt.team04.domain.match.review.dto.MatchRecordResponse;
-import com.kdt.team04.domain.match.review.repository.MatchRecordRepository;
+import com.kdt.team04.domain.matches.review.dto.MatchRecordResponse;
+import com.kdt.team04.domain.matches.review.repository.MatchRecordRepository;
 
 @Service
 public class MatchRecordGiverService {
