+++ src/main/java/com/kdt/team04/domain/matches/review/service/MatchReviewGiverService.java
@@ -1,9 +1,9 @@
-package com.kdt.team04.domain.match.review.service;
+package com.kdt.team04.domain.matches.review.service;
 
 import org.springframework.stereotype.Service;
 
-import com.kdt.team04.domain.match.review.dto.MatchReviewResponse;
-import com.kdt.team04.domain.match.review.repository.MatchReviewRepository;
+import com.kdt.team04.domain.matches.review.dto.MatchReviewResponse;
+import com.kdt.team04.domain.matches.review.repository.MatchReviewRepository;
 
 @Service
 public class MatchReviewGiverService {
