+++ src/test/java/com/kdt/team04/domain/team/service/TeamServiceTest.java
@@ -19,10 +19,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.kdt.team04.common.exception.EntityNotFoundException;
-import com.kdt.team04.domain.match.review.dto.MatchRecordResponse;
-import com.kdt.team04.domain.match.review.dto.MatchReviewResponse;
-import com.kdt.team04.domain.match.review.service.MatchRecordGiverService;
-import com.kdt.team04.domain.match.review.service.MatchReviewGiverService;
+import com.kdt.team04.domain.matches.review.dto.MatchRecordResponse;
+import com.kdt.team04.domain.matches.review.dto.MatchReviewResponse;
+import com.kdt.team04.domain.matches.review.service.MatchRecordGiverService;
+import com.kdt.team04.domain.matches.review.service.MatchReviewGiverService;
 import com.kdt.team04.domain.team.SportsCategory;
 import com.kdt.team04.domain.team.dto.TeamConverter;
 import com.kdt.team04.domain.team.dto.TeamRequest;
